[Ref] Introduced a new class, FlagSet, which makes working with flag sets easier. To use a flag set with this class, it has to be created with the FLAGSET macro.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@1327 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/Moddoc.h
@@ -260,8 +260,8 @@ public:
 	bool IsChannelSolo(CHANNELINDEX nChn) const;
 // -! NEW_FEATURE#0012
 	bool SurroundChannel(CHANNELINDEX nChn, bool bSurround);
-	bool SetChannelGlobalVolume(CHANNELINDEX nChn, UINT nVolume);
-	bool SetChannelDefaultPan(CHANNELINDEX nChn, UINT nPan);
+	bool SetChannelGlobalVolume(CHANNELINDEX nChn, uint16 nVolume);
+	bool SetChannelDefaultPan(CHANNELINDEX nChn, uint16 nPan);
 	bool IsChannelMuted(CHANNELINDEX nChn) const;
 	bool IsSampleMuted(SAMPLEINDEX nSample) const;
 	bool IsInstrumentMuted(INSTRUMENTINDEX nInstr) const;
